+++ elphick/geomet/utils/sampling.py
@@ -1,5 +1,0 @@
-import uuid
-
-
-def random_int():
-    return int(uuid.uuid4())